python-requirements.txt
azure-cli
azure.cli.core
pylint
